// src/DynamicDave.Umbraco.UrlInspector/Client/src/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: 'https://localhost:44414' | (string & {});
};

export type RedirectEntry = {
    statusCode: number;
    from: string;
    to?: string | null;
    culture?: string | null;
};

export type TestUrlRequest = {
    url: string;
};

export type TestUrlResponse = {
    allowed: boolean;
    statusCode?: number | null;
    location?: string | null;
    message?: string | null;
};

export type UrlEntry = {
    culture?: string | null;
    url?: string | null;
    status: string;
    message?: string | null;
};

export type UrlInspectorResponse = {
    urls: Array<UrlEntry>;
    incomingRedirects: Array<RedirectEntry>;
    outgoingRedirect?: RedirectEntry | null;
    canonical?: string | null;
};

export type GetUrlInspectorData = {
    body?: never;
    path: {
        key: string;
    };
    query?: never;
    url: '/umbraco/dynamicdave-urlinspector/api/v1/document/{key}';
};

export type GetUrlInspectorErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
    /**
     * Not Found
     */
    404: unknown;
};

export type GetUrlInspectorResponses = {
    /**
     * OK
     */
    200: UrlInspectorResponse;
};

export type GetUrlInspectorResponse = GetUrlInspectorResponses[keyof GetUrlInspectorResponses];

export type TestUrlData = {
    body?: TestUrlRequest;
    path?: never;
    query: {
        documentKey: string;
    };
    url: '/umbraco/dynamicdave-urlinspector/api/v1/test-url';
};

export type TestUrlErrors = {
    /**
     * Bad Request
     */
    400: unknown;
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
    /**
     * Not Found
     */
    404: unknown;
};

export type TestUrlResponses = {
    /**
     * OK
     */
    200: TestUrlResponse;
};

export type TestUrlResponse2 = TestUrlResponses[keyof TestUrlResponses];